requirements.txt

discord.py==2.4.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.4
PyYAML==6.0.2
rapidfuzz==3.9.6
orjson==3.10.7
